improved backup restore

+++ roles/openhab/templates/backup_restore/openhab.sh
@@ -1,12 +1,15 @@
 #!/usr/bin/bash
 
-echo "stop openhab"
-systemctl stop openhab
-
 FOLDER=`ls -td {{backup_path}}openhab/mapdb* 2>/dev/null | head -n 1`
-echo "backup folder: $FOLDER"
 
-cp -r $FOLDER {{global_lib}}openhab/persistance/mapdb/
+read -r -p "Are you sure to import '$FOLDER' into openhab mapdb? [y/N] " response
+if [[ "$response" =~ ^([yY][eE][sS]|[yY])$ ]]
+then
+    echo "stop openhab"
+    systemctl stop openhab
 
-echo "start openhab"
-systemctl start openhab
+    cp -r $FOLDER {{global_lib}}openhab/persistance/mapdb/
+
+    echo "start openhab"
+    systemctl start openhab
+fi
